Hive: Support column projection on Hive Profiles
Currently, Hive, HiveText and HiveRC profiles do not support column
projection. Add support for column projection for these profiles at the
PXF level. PXF can determine the following scenarios:

1. The number of columns of the file is higher than the number of
   columns defined at the Greenplum table definition.
2. The number of columns of the file is less than the number of columns
   defined at the Greenplum table definition.
3. A mixed number of columns is present in the files we are accessing.

For this to work, the column names on the Greenplum table definition
must match the column names on the Hive table definition, as the
projection is based on column names rather than column positions.